couple of fixes for converse api
