+++ src/bp/core/server.ts
@@ -7,7 +7,7 @@ import { UnlicensedError } from 'errors'
 import express from 'express'
 import rewrite from 'express-urlrewrite'
 import { createServer, Server } from 'http'
-import { inject, injectable, postConstruct, tagged } from 'inversify'
+import { inject, injectable, tagged } from 'inversify'
 import path from 'path'
 import portFinder from 'portfinder'
 
@@ -24,7 +24,7 @@ import ActionService from './services/action/action-service'
 import { AdminService } from './services/admin/service'
 import AuthService from './services/auth/auth-service'
 import { InvalidLicenseKey } from './services/auth/errors'
-import { CMS } from './services/cms/cms'
+import { CMSService } from './services/cms'
 import { ConverseService } from './services/converse'
 import { FlowService } from './services/dialog/flow/service'
 import { SkillService } from './services/dialog/skill/service'
@@ -32,8 +32,8 @@ import { LogsService } from './services/logs/service'
 import MediaService from './services/media'
 import { NotificationsService } from './services/notification/service'
 import { TYPES } from './types'
-const BASE_API_PATH = '/api/v1'
 
+const BASE_API_PATH = '/api/v1'
 const isProd = process.env.NODE_ENV === 'production'
 
 @injectable()
@@ -48,8 +48,7 @@ export default class HTTPServer {
   private readonly modulesRouter: ModulesRouter
   private readonly shortlinksRouter: ShortLinksRouter
   private readonly versioningRouter: VersioningRouter
-
-  private converseRouter: ConverseRouter | undefined
+  private readonly converseRouter: ConverseRouter
 
   constructor(
     @inject(TYPES.ConfigProvider) private configProvider: ConfigProvider,
@@ -58,7 +57,7 @@ export default class HTTPServer {
     private logger: Logger,
     @inject(TYPES.IsProduction) isProduction: boolean,
     @inject(TYPES.BotRepository) botRepository: BotRepository,
-    @inject(TYPES.CMS) cmsService: CMS,
+    @inject(TYPES.CMSService) cmsService: CMSService,
     @inject(TYPES.FlowService) flowService: FlowService,
     @inject(TYPES.ActionService) actionService: ActionService,
     @inject(TYPES.ModuleLoader) moduleLoader: ModuleLoader,
@@ -97,15 +96,10 @@ export default class HTTPServer {
     })
     this.contentRouter = new ContentRouter(this.adminService, this.authService, cmsService)
     this.versioningRouter = new VersioningRouter(this.adminService, this.authService, ghostService)
-
-    this.botsRouter.router.use('/content', this.contentRouter.router)
-    this.botsRouter.router.use('/versioning', this.versioningRouter.router)
-  }
-
-  @postConstruct()
-  async initAsync() {
     this.converseRouter = new ConverseRouter(this.converseService)
+    this.botsRouter.router.use('/content', this.contentRouter.router)
     this.botsRouter.router.use('/converse', this.converseRouter.router)
+    this.botsRouter.router.use('/versioning', this.versioningRouter.router)
   }
 
   resolveAsset = file => path.resolve(process.PROJECT_LOCATION, 'assets', file)
