+++ src/templates/data/global/content-types/builtin_carousel.js
@@ -3,30 +3,6 @@ const Card = require('./builtin_card')
 const url = require('url')
 
 function render(data) {
-  return {
-    text: ' ',
-    type: 'carousel',
-    elements: data.items.map(card => ({
-      title: card.title,
-      picture: card.image ? url.resolve(data.BOT_URL, card.image) : null,
-      subtitle: card.subtitle,
-      buttons: (card.actions || []).map(a => {
-        if (a.action === 'Say something') {
-          throw new Error('Webchat carousel does not support "Say something" action-buttons at the moment')
-        } else if (a.action === 'Open URL') {
-          return {
-            title: a.title,
-            url: a.url
-          }
-        } else {
-          throw new Error(`Webchat carousel does not support "${a.action}" action-buttons at the moment`)
-        }
-      })
-    }))
-  }
-}
-
-function renderForWeb(data) {
   const events = []
 
   if (data.typing) {
@@ -36,18 +12,35 @@ function renderForWeb(data) {
     })
   }
 
-  return [...events, render(data)]
-}
-
-function renderForApi(data) {
-  return [render(data)]
+  return [
+    ...events,
+    {
+      text: ' ',
+      type: 'carousel',
+      elements: data.items.map(card => ({
+        title: card.title,
+        picture: card.image ? url.resolve(data.BOT_URL, card.image) : null,
+        subtitle: card.subtitle,
+        buttons: (card.actions || []).map(a => {
+          if (a.action === 'Say something') {
+            throw new Error('Webchat carousel does not support "Say something" action-buttons at the moment')
+          } else if (a.action === 'Open URL') {
+            return {
+              title: a.title,
+              url: a.url
+            }
+          } else {
+            throw new Error(`Webchat carousel does not support "${a.action}" action-buttons at the moment`)
+          }
+        })
+      }))
+    }
+  ]
 }
 
 function renderElement(data, channel) {
-  if (channel === 'web') {
-    return renderForWeb(data)
-  } else if (channel === 'api') {
-    return renderForApi(data)
+  if (channel === 'web' || channel === 'api') {
+    return render(data)
   }
 
   return [] // TODO Handle channel not supported
