+++ src/templates/data/global/content-types/builtin_single_choice.js
@@ -1,18 +1,6 @@
 const base = require('./_base.js')
 
 function render(data) {
-  return {
-    on: 'webchat',
-    text: data.text,
-    quick_replies: data.choices.map(c => ({
-      title: c.title,
-      payload: c.value.toUpperCase()
-    })),
-    typing: data.typing
-  }
-}
-
-function renderForWeb(data) {
   const events = []
 
   if (data.typing) {
@@ -22,18 +10,23 @@ function renderForWeb(data) {
     })
   }
 
-  return [...events, render(data)]
-}
-
-function renderForApi(data) {
-  return [render(data)]
+  return [
+    ...events,
+    {
+      on: 'webchat',
+      text: data.text,
+      quick_replies: data.choices.map(c => ({
+        title: c.title,
+        payload: c.value.toUpperCase()
+      })),
+      typing: data.typing
+    }
+  ]
 }
 
 function renderElement(data, channel) {
-  if (channel === 'web') {
-    return renderForWeb(data)
-  } else if (channel === 'api') {
-    return renderForApi(data)
+  if (channel === 'web' || channel === 'api') {
+    return render(data)
   }
 
   return [] // TODO Handle channel not supported
