+++ src/templates/data/global/content-types/builtin_text.js
@@ -1,14 +1,6 @@
 const base = require('./_base.js')
 
 function render(data) {
-  return {
-    type: 'text',
-    markdown: true,
-    text: data.text
-  }
-}
-
-function renderForWeb(data) {
   const events = []
 
   if (data.typing) {
@@ -18,18 +10,19 @@ function renderForWeb(data) {
     })
   }
 
-  return [...events, render(data)]
-}
-
-function renderForApi(data) {
-  return [render(data)]
+  return [
+    ...events,
+    {
+      type: 'text',
+      markdown: true,
+      text: data.text
+    }
+  ]
 }
 
 function renderElement(data, channel) {
-  if (channel === 'web') {
-    return renderForWeb(data)
-  } else if (channel === 'api') {
-    return renderForApi(data)
+  if (channel === 'web' || channel === 'api') {
+    return render(data)
   }
 
   return [] // TODO
